hibernate

+++ src/main/java/org/homework/product_dao/ProductDao.java
@@ -12,54 +12,67 @@ import java.util.List;
 public class ProductDao {
     private final EntityManagerFactory EMFactory;
 
-
     public ProductDao(EntityManagerFactory EMFactory) {
         this.EMFactory = EMFactory;
     }
 
 
     public Product findOneById(Long id) {
+
         EntityManager em = EMFactory.createEntityManager();
         em.getTransaction().begin();
-        Product product = em.find(Product.class, id);
 
+        Product product = em.find(Product.class, id);
         em.getTransaction().commit();
 
         em.close();
+        if (product == null) {
+            System.out.println("Нет такого продукта!");
+        }
+
         return product;
     }
 
 
     public List<Product> findAll() {
+
         EntityManager em = EMFactory.createEntityManager();
         em.getTransaction().begin();
+
         List<Product> from_products = em.createQuery("from Product ", Product.class).getResultList();
+
         em.getTransaction().commit();
         em.close();
+
         return from_products;
     }
 
     public void deleteById(Long id) {
+
         EntityManager em = EMFactory.createEntityManager();
         em.getTransaction().begin();
-        em.remove(em.find(Product.class, id));
+
+        if (em.find(Product.class, id) != null) {
+            em.remove(em.find(Product.class, id));
+        } else {
+            System.out.println("Нет такого продкукта!");
+        }
+
         em.getTransaction().commit();
         em.close();
 
     }
 
     public Product saveOrUpdate(Product product) {
+
         EntityManager em = EMFactory.createEntityManager();
         em.getTransaction().begin();
-        List<Product> productList = em.createQuery("from Product ", Product.class).getResultList();
-        if (productList.contains(product)) {
-            product.setTitle(product.getTitle());
-            product.setPrice(product.getPrice());
-        } else {
-            em.merge(product);
-        }
+
+        em.merge(product);
+
         em.getTransaction().commit();
         em.close();
+
         return product;
     }
